refactor : 메서드 순서를 public -> private 으로 통일, 테스트코드 위치 변경
