+++ src/test/java/model/LottoNumbersTest.java
@@ -9,6 +9,28 @@ import org.junit.jupiter.api.Test;
 
 class LottoNumbersTest {
     @Test
+    @DisplayName("로또 번호 개수 테스트")
+    public void lottoNumbersCountTest() {
+        // given & when
+        LottoNumbers lottoNumbers = new LottoNumbers();
+        List<LottoNumber> numbers = lottoNumbers.getNumbers();
+        // then
+        assertThat(numbers.size()).isEqualTo(LottoNumbers.LOTTO_NUMBER_COUNT);
+    }
+
+    @Test
+    @DisplayName("로또 번호 범위 테스트")
+    public void lottoNumberBoundTest() {
+        // given & when
+        LottoNumbers lottoNumbers = new LottoNumbers();
+        List<LottoNumber> numbers = lottoNumbers.getNumbers();
+        // then
+        for (LottoNumber number : numbers) {
+            assertThat(number.getNumber()).isBetween(1, 45);
+        }
+    }
+
+    @Test
     @DisplayName("로또번호 매칭 테스트")
     public void countMatchNumberTest() {
         // given
