+++ src/test/java/model/LottoWinRankTest.java
@@ -1,11 +1,8 @@
-package constant;
+package model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import model.LottoNumbers;
-import model.LottoWinRank;
-import model.WinLotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
